+++ test/test_helper.exs
@@ -4,15 +4,15 @@ defmodule Scrivener.TestCase do
   using(opts) do
     quote do
       use ExSpec, unquote(opts)
-      import Ecto.Query
+      import EctoOne.Query
     end
   end
 
   setup do
-    Ecto.Adapters.SQL.begin_test_transaction(Scrivener.Repo)
+    EctoOne.Adapters.SQL.begin_test_transaction(Scrivener.Repo)
 
     ExUnit.Callbacks.on_exit(fn ->
-      Ecto.Adapters.SQL.rollback_test_transaction(Scrivener.Repo)
+      EctoOne.Adapters.SQL.rollback_test_transaction(Scrivener.Repo)
     end)
   end
 end